backend/requirements.txt
# FastAPI and dependencies
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

# Database
sqlalchemy
psycopg2-binary
alembic

# Redis
redis==5.0.1
celery==5.3.4

# Search
elasticsearch==8.11.0

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI
openai==1.3.7
langchain==0.0.348

# XML processing
lxml==4.9.3
xmlschema==2.5.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography>=42.0.0
bcrypt>=4.0.0

# Utilities
python-slugify==8.0.1
Pillow==10.1.0
pyyaml==6.0.1